feat(page-types): define TCompareStats, TCompareChartPoint, and TDiffStats types for climate comparison
